Sum of any number of integers
The algorithm can now calculate the sum of any given number of integers in the form of a continuous string with single spaces between them. For the moment, in case of non numeric input, the calculator doesn't resolve the issue nor does it display a message. It just waits for a valid input. There is the option of two keywords: /exit and /help which end the program and print a help message respectively.